requirements.txt
google-adk==1.19.0
a2a-sdk==0.3.16
gradio==6.0.1
python-dotenv>=1.2.1
./mcp_server